shared/pyproject.toml
[project]
name = "naql-common"
version = "0.1.0"
description = "Shared libraries for Naql.ai microservices"
requires-python = ">=3.12"

dependencies = [
    "pydantic>=2.6,<3",
    "sqlalchemy>=2.0,<3",
    "asyncpg>=0.29,<1",
    "redis[hiredis]>=5.0,<6",
    "nats-py>=2.7,<3",
    "pyjwt>=2.8,<3",
    "h3>=4.0,<5",
    "uuid7>=0.1,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.3,<1",
    "mypy>=1.9,<2",
]
